refactor: broken MT5PasswordModal into smaller component
